EXP modified plotting file for the expes
